+++ lib/support.c
@@ -159,11 +159,6 @@ X(i64)
         CAMLreturn(namespace##_wrap(C_INT_MAX(namespace)));                    \
     }                                                                          \
                                                                                \
-    value checked_oint_##namespace##_of_int_unchecked(value x) {               \
-        CAMLparam1(x);                                                         \
-        CAMLreturn(namespace##_wrap((C_INT_TY(namespace))Int_val(x)));         \
-    }                                                                          \
-                                                                               \
     value checked_oint_##namespace##_print(value x) {                          \
         CAMLparam1(x);                                                         \
                                                                                \
@@ -217,3 +212,29 @@ X(i128)
 #undef X
 #undef X_OP1
 #undef X_OP2
+
+#define X(signedness, namespace)                                               \
+    value checked_oint_##signedness##128##_of_##namespace##_unchecked(         \
+        value x) {                                                             \
+        CAMLparam1(x);                                                         \
+        CAMLreturn(signedness##128##_wrap(                                     \
+            (C_INT_TY(namespace))C_VALUE(namespace, x)));                      \
+    }
+
+X(u, u64)
+X(i, i64)
+
+#undef X
+
+#define X(namespace)                                                           \
+    value checked_oint_i128_to_##namespace(value x) {                          \
+        CAMLparam1(x);                                                         \
+        CAMLreturn(                                                            \
+            OCAML_VALUE(namespace, (C_INT_TY(namespace))C_VALUE(i128, x)));    \
+    }
+
+X(int)
+X(i32)
+X(i64)
+
+#undef X